requirements.txt
# ===========================
# PG Accountant - Requirements
# ===========================

# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0

# LangGraph & LangChain
langgraph==0.1.4
langchain==0.2.3
langchain-anthropic==0.1.11
langchain-core==0.2.5

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9   # APScheduler SQLAlchemyJobStore (sync PostgreSQL driver)

# PDF Parsing
pdfplumber==0.11.0
pymupdf==1.24.3
camelot-py[cv]==0.11.0

# Data Processing
pandas==2.2.2
openpyxl==3.1.2
xlsxwriter==3.2.0
numpy==1.26.4

# HTTP & Async
httpx==0.27.0
aiohttp==3.9.5
aiofiles==23.2.1

# Scheduling & Background Tasks
apscheduler==3.10.4
celery==5.4.0

# CLI
click==8.1.7
rich==13.7.1
questionary==2.0.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
dateparser==1.2.0
arrow==1.3.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Logging
loguru==0.7.2

# Security
cryptography==42.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WhatsApp (Meta Cloud API — uses httpx, already listed above)

# n8n/Webhook
requests==2.32.3
websockets==12.0
